requirements.txt
opencv-python
pytesseract
pydirectinput-rgx
mss
pillow
pynput
pyside6